+++ R/x.import.R
@@ -26,7 +26,7 @@
 #' @param alias the alias object as created by the \link{import_as} function. \cr
 #' @param ... further arguments to be passed to \link[utils]{help}.
 #' @param which The attributes to list. If \code{NULL}, all attributes will be returned. \cr
-#' Possibilities: "pkgs", "conflicts", "versions", "args", and "ordered_object_names".
+#' Possibilities: "pkgs", "conflicts", "args", and "ordered_object_names".
 #' @param x the object/function to be tested.
 #'
 #' @details
@@ -80,10 +80,6 @@
 #' in case of conflicts. \cr
 #' The "conflicts" attribute returns a data.frame showing exactly which functions overwrite
 #' functions from earlier named packages, and as such "win" the conflicts. \cr
-#' \cr
-#' For \code{attr.import(alias, which = "versions")}: \cr
-#' A data.frame, giving the version of every package imported under the alias,
-#' ignoring re-exports. \cr
 #' \cr
 #' For \code{attr.import(alias, which = "args")}: \cr
 #' Returns a list of input arguments.
@@ -246,7 +242,7 @@ attr.import <- function(alias, which = NULL) {
     return(alias$.__attributes__.)
   }
 
-  allowed_which <- c("pkgs", "conflicts", "versions", "args", "ordered_object_names")
+  allowed_which <- c("pkgs", "conflicts", "args", "ordered_object_names")
   if(!isTRUE(which %in% allowed_which)) {
     stop("unknown `which` given")
   }